cmd/compile: refactor shared equality
This patch moves some of the shared equality checks into their
own package.
